Universe inits to 1000x1000 grid with pattern from file in the center. Also changed some cargo clippy suggestions

+++ src/lib.rs
@@ -99,7 +99,7 @@ impl Universe {
         Universe { width, height, cells }
     }
 
-    pub fn run_iterations(&mut self, iterations: u32) {
+    pub fn run_iterations(&mut self, iterations: usize) {
         print_memory_usage("Before Running Iterations");
 
         for i in 0..iterations {
